Handle SSH_MSG_GLOBAL_REQUEST; build EBOOT.PBP via EXTRA_TARGETS
- CI trace decoded: OpenSSH sends 'hostkeys-00@openssh.com' global
  request right after auth success (want_reply=1). Reply
  REQUEST_FAILURE per RFC 4254 §4 in expect_msg and the data pump
  (previously 'session failed' after auth)
- Pack EBOOT.PBP: EXTRA_TARGETS+PSP_EBOOT_TITLE so build.mak emits
  it; workflow checks EBOOT.PBP at root, not PSPSH/EBOOT.PBP
- net_psp.c: use inet_addr/sockaddr_in (SDK has no
  sceNetInetInetAddr; connect takes (fd, sa, len))
- fix %lu format warning in transport debug

+++ src/ssh/transport.c
@@ -244,7 +244,7 @@ int sshe_tx_recv_packet(sshe_tx *t, sshe_buf *out)
     /* payload = bytes after [len(4)][padlen(1)] */
     pad_len = first[4];
     if (pad_len == 0 || pad_len > 255 || packet_len < 1 + pad_len) {
-        DBG("recv: bad pad_len %u\n", pad_len);
+        DBG("recv: bad pad_len %lu\n", (unsigned long)pad_len);
         free(body);
         return -1;
     }
